Task list polished a bit Big refactor to organize the code
