+++ Timetracker/Delegates/AppDelegate+NSMenuDelegate.swift
