+++ Timetracker/Delegates/AppDelegate+TaskPingReceiver.swift
