+++ Timetracker/Managers/DockIconManager.swift
