--- Timetracker/TaskRunnerWindowController.swift
@@ -1,14 +1,0 @@
-//
-//  TaskRunnerWindowController.swift
-//  Timetracker
-//
-//  Created by Antonio Ribeiro on 05/05/16.
-//  Copyright © 2016 Antonio Ribeiro. All rights reserved.
-//
-
-import Foundation
-import Cocoa
-
-class TaskRunnerWindowController: NSWindowController {
-
-}
