+++ Timetracker/Views/ManualTaskViewController.swift
