+++ Timetracker/copyright.txt
@@ -7,3 +7,4 @@ FRAMEWORKS:
 
 SwiftyBeaver (https://github.com/SwiftyBeaver/SwiftyBeaver)
 Preferences (https://github.com/sindresorhus/Preferences)
+SwiftDate (https://github.com/malcommac/SwiftDate)
